LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/stylenonce-blocked.sub-expected.txt

CONSOLE MESSAGE: Refused to apply a stylesheet because its hash, its nonce, or 'unsafe-inline' does not appear in the style-src directive of the Content Security Policy.
This text should be green.

Style that does not match a 'nonce-*' expression in 'style-src' should not be applied to the page.


Harness Error (TIMEOUT), message = null

NOTRUN Should fire securitypolicyviolation
PASS stylenonce-blocked
